pyproject.toml
[project]
name = "douscrapper"
version = "0.1.0"
description = "Automated daily monitoring of Diário Oficial da União (DOU)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "structlog>=24.1.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "responses>=0.25.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0",
    "types-pyyaml>=6.0.0",
    "types-pytz>=2024.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
# E: pycodestyle errors
# F: pyflakes
# B: bugbear
# I: isort

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
